first playable version

+++ core/src/pl/swidurski/pacman/map/elements/Eatable.java
@@ -4,7 +4,7 @@ package pl.swidurski.pacman.map.elements;
  * Created by Krystek on 2016-04-11.
  */
 public interface Eatable {
-    double getPoints();
+    int getPoints();
 
     boolean isEatable();
 }
